Move academy.agent to academy.runner
